SF-6 Refactored profile, it now is exclusively a part of user
