+++ src/components/navbar/NavBarHeader.vue
@@ -86,7 +86,7 @@ import { useRouter } from "vue-router";
 import { onMounted, reactive, ref, watch } from 'vue'
 import { storeToRefs } from 'pinia'
 import User from '@/functional_components/API/user/user.class'
-import { Profile } from '@/functional_components/API/profile/profile.class'
+import { Profile } from '@/functional_components/API/user/profile.class'
 
 const apiStore = useApiStore();
 const router = useRouter();
