+++ src/functional_components/API/user/user.service.ts
@@ -2,7 +2,7 @@ import BaseService from '../abstract.service'
 import User from './user.class'
 import API from '@/functional_components/API/api'
 import { CreateUserDto } from '@/functional_components/API/user/dto/create-user.dto'
-import { Profile } from '@/functional_components/API/profile/profile.class'
+import { Profile } from '@/functional_components/API/user/profile.class'
 import { plainToClass, plainToInstance } from 'class-transformer'
 
 export class UserService extends BaseService<User> {
